requirements.txt
Flask
Flask-SQLAlchemy
pyjwt
psycopg2